Refactor react components to be more organized.
